refactor(logging): split up logging logic
This doesn't make any changes beyond reorganizing the logging files by splitting them up between individaul logger files (mcpLogger, baseLogger...), types and definitions, all re-exported through logging/index.js

This both prepares us for the package-split futuree and provides a better basis for future refactoring of the logging setup.
