+++ src/common/atlas/apiClient.ts
@@ -4,7 +4,7 @@ import { ApiClientError } from "./apiClientError.js";
 import type { paths, operations } from "./openapi.js";
 import type { CommonProperties, TelemetryEvent } from "../../telemetry/types.js";
 import { packageInfo } from "../packageInfo.js";
-import type { LoggerBase } from "../logger.js";
+import type { LoggerBase } from "../logging/index.js";
 import { createFetch } from "@mongodb-js/devtools-proxy-support";
 import { Request as NodeFetchRequest } from "node-fetch";
 import type { Credentials, AuthProvider } from "./auth/authProvider.js";
