+++ src/common/atlas/cluster.ts
@@ -4,7 +4,7 @@ import type {
     FlexClusterDescription20241113,
 } from "./openapi.js";
 import type { ApiClient } from "./apiClient.js";
-import { LogId } from "../logger.js";
+import { LogId } from "../logging/index.js";
 import { ConnectionString } from "mongodb-connection-string-url";
 
 type AtlasProcessId = `${string}:${number}`;
