+++ src/common/connectionManager.ts
@@ -4,7 +4,7 @@ import { generateConnectionInfoFromCliArgs, type ConnectionInfo } from "@mongosh
 import type { DeviceId } from "../helpers/deviceId.js";
 import { type UserConfig } from "./config/userConfig.js";
 import { MongoDBError, ErrorCodes } from "./errors.js";
-import { type LoggerBase, LogId } from "./logger.js";
+import { type LoggerBase, LogId } from "./logging/index.js";
 import { packageInfo } from "./packageInfo.js";
 import { type AppNameComponents, setAppNameParamIfMissing } from "../helpers/connectionOptions.js";
 import {
