+++ src/common/logging/loggingTypes.ts
@@ -1,0 +1,19 @@
+import type { MongoLogId } from "mongodb-log-writer";
+
+import type { LoggingMessageNotification } from "@modelcontextprotocol/sdk/types.js";
+
+export type LogLevel = LoggingMessageNotification["params"]["level"];
+
+export interface LogPayload {
+    id: MongoLogId;
+    context: string;
+    message: string;
+    noRedaction?: boolean | LoggerType | LoggerType[];
+    attributes?: Record<string, string>;
+}
+
+export type LoggerType = "console" | "disk" | "mcp";
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type EventMap<T> = Record<keyof T, any[]> | DefaultEventMap;
+export type DefaultEventMap = [never];
