+++ src/lib.ts
@@ -25,7 +25,13 @@ export { createMCPConnectionManager, defaultCreateConnectionManager };
 export { defaultCreateApiClient } from "./common/atlas/apiClient.js";
 export { defaultCreateAtlasLocalClient } from "./common/atlasLocal.js";
 
-export { LoggerBase, type LogPayload, type LoggerType, type LogLevel, CompositeLogger } from "./common/logger.js";
+export {
+    LoggerBase,
+    type LogPayload,
+    type LoggerType,
+    type LogLevel,
+    CompositeLogger,
+} from "./common/logging/index.js";
 export { StreamableHttpRunner } from "./transports/streamableHttp.js";
 export { StdioRunner } from "./transports/stdio.js";
 export {
