+++ src/server.ts
@@ -2,8 +2,8 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import type { Session } from "./common/session.js";
 import type { Transport } from "@modelcontextprotocol/sdk/shared/transport.js";
 import { Resources } from "./resources/resources.js";
-import type { LogLevel } from "./common/logger.js";
-import { LogId, McpLogger } from "./common/logger.js";
+import type { LogLevel } from "./common/logging/index.js";
+import { LogId, McpLogger } from "./common/logging/index.js";
 import type { Telemetry } from "./telemetry/telemetry.js";
 import type { UserConfig } from "./common/config/userConfig.js";
 import { type ServerEvent } from "./telemetry/types.js";
