+++ src/telemetry/telemetry.ts
@@ -1,7 +1,7 @@
 import type { Session } from "../common/session.js";
 import type { BaseEvent, CommonProperties } from "./types.js";
 import type { UserConfig } from "../common/config/userConfig.js";
-import { LogId } from "../common/logger.js";
+import { LogId } from "../common/logging/index.js";
 import type { ApiClient } from "../common/atlas/apiClient.js";
 import { MACHINE_METADATA } from "./constants.js";
 import { EventCache } from "./eventCache.js";
