+++ src/tools/assistant/listKnowledgeSources.ts
@@ -1,7 +1,7 @@
 import type { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
 import { formatUntrustedData, type OperationType, type ToolCategory } from "../tool.js";
 import { AssistantToolBase } from "./assistantTool.js";
-import { LogId } from "../../common/logger.js";
+import { LogId } from "../../common/logging/index.js";
 import { stringify as yamlStringify } from "yaml";
 
 export type KnowledgeSource = {
