+++ src/transports/base.ts
@@ -4,8 +4,8 @@ import { type AnyToolClass, Server, type ServerOptions } from "../server.js";
 import { Session, type SessionOptions } from "../common/session.js";
 import { Telemetry } from "../telemetry/telemetry.js";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import type { LoggerBase } from "../common/logger.js";
-import { CompositeLogger, ConsoleLogger, DiskLogger, McpLogger } from "../common/logger.js";
+import type { LoggerBase } from "../common/logging/index.js";
+import { CompositeLogger, ConsoleLogger, DiskLogger, McpLogger } from "../common/logging/index.js";
 import { ExportsManager } from "../common/exportsManager.js";
 import { DeviceId } from "../helpers/deviceId.js";
 import { Keychain } from "../common/keychain.js";
