+++ src/transports/streamableHttp.ts
@@ -2,8 +2,8 @@ import express from "express";
 import type http from "http";
 import type { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
 import { isInitializeRequest } from "@modelcontextprotocol/sdk/types.js";
-import type { LoggerBase } from "../common/logger.js";
-import { CompositeLogger, LogId } from "../common/logger.js";
+import type { LoggerBase } from "../common/logging/index.js";
+import { CompositeLogger, LogId } from "../common/logging/index.js";
 import { SessionStore } from "../common/sessionStore.js";
 import {
     TransportRunnerBase,
