+++ tests/integration/telemetry.test.ts
@@ -2,7 +2,7 @@ import { Telemetry } from "../../src/telemetry/telemetry.js";
 import { Session } from "../../src/common/session.js";
 import { DeviceId } from "../../src/helpers/deviceId.js";
 import { describe, expect, it } from "vitest";
-import { CompositeLogger } from "../../src/common/logger.js";
+import { CompositeLogger } from "../../src/common/logging/index.js";
 import { MCPConnectionManager } from "../../src/common/connectionManager.js";
 import { connectionErrorHandler } from "../../src/common/connectionErrorHandler.js";
 import { ExportsManager } from "../../src/common/exportsManager.js";
