+++ tests/unit/helpers/connectionOptions.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { setAppNameParamIfMissing } from "../../../src/helpers/connectionOptions.js";
 import { DeviceId } from "../../../src/helpers/deviceId.js";
-import { CompositeLogger } from "../../../src/common/logger.js";
+import { CompositeLogger } from "../../../src/common/logging/index.js";
 
 const MockDeviceId = vi.mocked(DeviceId.create(new CompositeLogger()));
