+++ tests/unit/logger.test.ts
@@ -1,7 +1,7 @@
 import type { MockInstance } from "vitest";
 import { describe, beforeEach, afterEach, vi, it, expect } from "vitest";
-import type { LoggerType, LogLevel } from "../../src/common/logger.js";
-import { CompositeLogger, ConsoleLogger, DiskLogger, LogId, McpLogger } from "../../src/common/logger.js";
+import type { LoggerType, LogLevel } from "../../src/common/logging/index.js";
+import { CompositeLogger, ConsoleLogger, DiskLogger, LogId, McpLogger } from "../../src/common/logging/index.js";
 import os from "os";
 import * as path from "path";
 import * as fs from "fs/promises";
